[Refactor] cafe.rb が肥大化していたのでファイルを分割した
